Parrying implemented + color-coded balls.
You can now parry a ball if you are holding a ball.  The parried ball switches teams and gains a ludicrous  amount of speed.
Catching and parrying can be done with the left trigger on a controller or the E key on the keyboard.  Catching with mouseclick or the right trigger has been removed.
Added new boolean to Player called armorlock that makes the Player permanently parry for testing purposes.
Dummy and Players are now placed on the ground instead of the air.
Added an enemy Player to LevelZero that has armorlock on for testing purposes. LevelZero also spawns a ball from this player.
Color coded ball projectiles have been added, however there are only two colors.  Even team ids result in a green outline around the ball.  Odd numbers result in a red outline.